+++ cartridge/confapplier.lua
@@ -190,16 +190,6 @@ local function validate_config(clusterwide_config, _)
         conf_old = {}
     end
 
-    if vars.state == 'Unconfigured' then
-        local ok, err = topology.validate_self_uri(
-            clusterwide_config:get_readonly('topology')
-        )
-
-        if not ok then
-            return nil, err
-        end
-    end
-
     local ok, err = ddl_manager.validate_config(conf_new, conf_old)
     if not ok then
         return nil, err