+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Fixed
 - Eventual failover may miss an event while roles are being reconfigured.
 - Compatibility with pipe logging
   (see: https://github.com/tarantool/tarantool/issues/5220)
+- Invalid state transition from Unconfigured to InitError
 
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Enhanced is WebUI